fix(review): address CI + PR review feedback
- SonarCloud duplication: extract shared scripts/lib/profile-override-runner.ts
  and make both cleanup-new-kennel-profiles and cleanup-cycle6-profile-overrides
  thin wrappers (mirrors backfill-runner.ts factoring) — removes the new-code
  duplication that failed the quality gate.
- parseBaliTitle (#1838): guard the trailing-date strip with VALID_MONTHS +
  a trailing-only check so a location containing a date-like token can't be
  truncated (Gemini + claude-review).
- cleanup script: set process.exitCode instead of process.exit() so the
  runner's prisma.$disconnect()/pool.end() drain cleanly (Gemini).
- phoenixhhh.test.ts: fix a pre-existing month-boundary flake — route the
  fetch mock by URL (month-list vs ?event= detail) so it's date-stable.
